Move class CartDeliveryOption, fix typo, add test for hideDiscount
